+++ software/distro/setup/planktoscope-app-env/node-red-frontend/install.sh
@@ -8,6 +8,10 @@ repo_root=$(dirname $(dirname $(dirname $distro_setup_files_root)))
 
 # Get command-line args
 hardware_type="$1" # should be either adafruithat or planktoscopehat
+if [ $hardware_type = "segmenter-only" ]; then
+  echo "Warning: setting up adafruithat version of Node-RED dashboard for hardware type: $hardware_type"
+  hardware_type=adafruithat
+fi
 
 # Install dependencies
 # smbus is needed by some python3 nodes in the Node-RED dashboard for the Adafruit HAT.
@@ -15,33 +19,30 @@ hardware_type="$1" # should be either adafruithat or planktoscopehat
 # default, we do `pip3 install` as the pi user. This makes the smbus2 module available to Node-RED.
 # FIXME: get rid of the Node-RED nodes depending on smbus! That functionality should be moved into
 # the Python backend.
-pip3 install smbus2==0.4.3
+# Note: for bookworm we need to install the apt package; for bullseye there is no apt package
+if ! sudo apt-get install -y python3-smbus2 ; then
+  sudo apt-get install -y python3-pip
+  pip3 install smbus2==0.4.3
+fi
 
 # Install Node-RED
 # TODO: run Node-RED in a Docker container instead
 curl -sL https://raw.githubusercontent.com/node-red/linux-installers/master/deb/update-nodejs-and-nodered \
   | bash -s - --confirm-install --confirm-pi --no-init
 
-# Add systemd service modification to make Node-RED wait until Mosquitto has started
-# FIXME: The Node-RED frontend should instead be fixed so that it does not need to wait until
-# Mosquitto has started in order to work - until Mosquitto is up, the frontend should display a
-# message to the user and ask them to wait.
-sudo mkdir -p /etc/systemd/system/nodered.service.d
-file="/etc/systemd/system/nodered.service.d/override.conf"
-sudo cp "$config_files_root$file" "$file"
-sudo systemctl daemon-reload
-sudo systemctl enable nodered.service
-
 # Select the enabled dashboard
+mkdir -p $HOME/.node-red
 cp "$repo_root/software/node-red-dashboard/flows/$hardware_type.json" \
   $HOME/.node-red/flows.json
+mkdir -p $HOME/PlanktoScope
 cp "$repo_root/software/node-red-dashboard/default-configs/$hardware_type-latest.config.json" \
   $HOME/PlanktoScope/config.json
 
 # Copy required dependencies with hard-coded paths in the Node-RED dashboard
-# TODO: get rid of this when we move the Node-RED dashboard out to its own repository
+# TODO: get rid of this when we remove usb_backup.sh
 mkdir -p $HOME/PlanktoScope/scripts
 directory="scripts/bash"
+# TODO: get rid of this when we move the Node-RED dashboard out to its own repository
 cp -r "$repo_root/$directory" $HOME/PlanktoScope/$directory
 mkdir -p $HOME/PlanktoScope/software/node-red-dashboard
 directory="software/node-red-dashboard/default-configs"